Minor refactoring. Only thing left now is the actual simulation.
