+++ src/bear_main.h
@@ -5,8 +5,10 @@
 #include "glad.h"
 
 #include "bear_types.h"
+#include "bear_array.h"
 #include "audio/bear_audio.h"
 #include "ecs/bear_ecs.h"
+#include "physics/bear_physics.h"
 
 #define WINDOW_WIDTH 800
 #define WINDOW_HEIGHT 600
@@ -71,6 +73,9 @@ struct World
 		Vec3f position;
 	} camera;
 
+	// Physics
+	Physics phy;
+
 	// A clock for timing.
 	CLK clk;
 
@@ -106,5 +111,5 @@ struct World
 
 #include "ecs/bear_ecs_init.cpp"
 
-#include "bear_array.h"
+#include "bear_array.cpp"
 #include "bear_input.h"
